+++ teacher-service/src/main/java/org/aston/ems/teacher_service/dao/api/ITaskRepository.java
@@ -1,5 +1,6 @@
-package org.aston.ems.teacher_service.dao;
+package org.aston.ems.teacher_service.dao.api;
 
+import org.aston.ems.teacher_service.dao.model.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
