+++ teacher-service/src/main/java/org/aston/ems/teacher_service/dao/model/BaseEntity.java
@@ -1,4 +1,4 @@
-package org.aston.ems.teacher_service.dao;
+package org.aston.ems.teacher_service.dao.model;
 
 import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
@@ -7,12 +7,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.io.Serializable;
+
 @MappedSuperclass
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public abstract class BaseEntity {
+public abstract class BaseEntity implements Serializable {
     @Id
     private long id;
 }
