+++ teacher-service/src/main/java/org/aston/ems/teacher_service/dao/model/Task.java
@@ -1,11 +1,9 @@
-package org.aston.ems.teacher_service.dao;
+package org.aston.ems.teacher_service.dao.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
+import org.aston.ems.teacher_service.dao.model.BaseEntity;
 
 @Getter
 @Setter
@@ -16,7 +14,8 @@ import lombok.Setter;
 public class Task extends BaseEntity {
     private long studentId;
     private long teacherId;
-    private String task;
+    private String content;
+    private String answer;
     private boolean checked;
     private int mark;
 }
